IDP_AI_Pipeline/requirements.txt
numpy
google-genai
python-dotenv
faiss-cpu
